first

+++ src/app/seller-home/seller-home.component.ts
@@ -1,8 +1,16 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {SellerProductService} from "../services/seller-product.service";
 import {Product} from "../data-types";
 import {MatIconModule} from "@angular/material/icon";
 import {NgForOf, NgOptimizedImage} from "@angular/common";
+import {MatTableDataSource, MatTableModule} from "@angular/material/table";
+import {MatSort, MatSortModule} from "@angular/material/sort";
+import {MatButtonModule} from "@angular/material/button";
+import {MatDialog} from "@angular/material/dialog";
+import {
+  SellerProductDeleteDialogComponent
+} from "../seller-product-delete-dialog/seller-product-delete-dialog.component";
+import {RouterLink} from "@angular/router";
 
 @Component({
   selector: 'app-seller-home',
@@ -10,36 +18,60 @@ import {NgForOf, NgOptimizedImage} from "@angular/common";
   imports: [
     MatIconModule,
     NgForOf,
-    NgOptimizedImage
+    NgOptimizedImage,
+    MatTableModule,
+    MatSortModule,
+    MatButtonModule,
+    RouterLink,
   ],
   templateUrl: './seller-home.component.html',
   styleUrl: './seller-home.component.css',
   providers:[SellerProductService]
 })
 export class SellerHomeComponent implements OnInit{
-  productList:undefined|Product[];
+
+  productList:Product[]|undefined;
   productMessage: undefined|string;
+  displayedColumns: string[] = ['image', 'name', 'price', 'color', 'category', 'description', 'actions'];
+  dataSource:any;
+
+  @ViewChild(MatSort, {static: true}) sort!: MatSort;
   ngOnInit(): void {
     this.list();
+
   }
-  constructor(private product:SellerProductService) {
+  constructor(private product:SellerProductService, private dialog:MatDialog) {
   }
   deleteProduct(id: number) {
-    this.product.deleteProduct(id).subscribe((result) => {
-      if (result) {
-        this.productMessage = 'Product '+id+' is deleted';
+    const dialogRef=this.dialog.open(SellerProductDeleteDialogComponent,{
+      width:'400px',
+      data:{id}
+    })
+    dialogRef.afterClosed().subscribe(res=>{
+      if(res==='Yes')
+      {
+        this.product.deleteProduct(id).subscribe((result) => {
+          if (result) {
+            this.productMessage = 'Product is deleted';
 
-        this.list();
+            this.list();
+          }
+        });
+        setTimeout(() => {
+          this.productMessage = undefined;
+        }, 3000);
       }
-    });
-    setTimeout(() => {
-      this.productMessage = undefined;
-    }, 3000);
+    })
+
   }
   list(){
     this.product.productList().subscribe((result)=>{
-      console.warn(result);
-      this.productList=result
+      this.productList =result ;
+      this.dataSource=new MatTableDataSource<Product>(this.productList);
+
     })
   }
+
+
+
 }
